requirements.txt
openpyxl==3.0.10
opensearch-py==2.0.0
boto3==1.24.89
elasticsearch==7.17.0
pandas==2.0.2
python-dotenv==1.0.0